feat(goals): mount the banner dock in the shell, retire the in-content strips
The goal-agent voices move out of the day/habits pages into the
shell-level dock (WP2 part 2): one rotating slot docked at the bottom of
the content region on desktop and above the bottom nav on mobile, on the
main working tabs only (Tasks, DailyOS, Habits — Settings and Logbook
excluded per the handover). It collapses when no goal is speaking, so it
costs no space until it has something to say.

The dead GoalBannerStrip widget is removed; its still-shared exposure
tracker is extracted to goal_banner_exposure_tracker.dart and its tests
ported to a minimal host. The day page and habits tab drop their strip
mounts. CHANGELOG and flatpak metainfo updated to describe the dock (the
feature is unreleased and flag-gated, so the existing entry is corrected
rather than a new one added).
